--- a/Gems/SurfaceData/Code/Source/Components/SurfaceDataColliderComponent.h
+++ b/Gems/SurfaceData/Code/Source/Components/SurfaceDataColliderComponent.h
@@ -12,6 +12,7 @@
 #include <AzCore/Component/Component.h>
 #include <AzCore/Component/TickBus.h>
 #include <AzCore/Component/TransformBus.h>
+#include <AzCore/std/parallel/shared_mutex.h>
 #include <AzFramework/Physics/ColliderComponentBus.h>
 
 #include <SurfaceData/SurfaceDataTypes.h>
@@ -98,7 +99,8 @@ namespace SurfaceData
 
         // cached data
         AZStd::atomic_bool m_refresh{ false };
-        mutable AZStd::recursive_mutex m_cacheMutex;
+        mutable AZStd::shared_mutex m_cacheMutex;
         AZ::Aabb m_colliderBounds = AZ::Aabb::CreateNull();
+        SurfaceTagWeightMap m_newPointWeights;
     };
 }