… Gems/SurfaceData/Code/Source/SurfaceDataSystemComponent.h
… Gems/SurfaceData/Code/Tests/SurfaceDataColliderComponentTest.cpp
… Gems/SurfaceData/Code/Tests/SurfaceDataTest.cpp
… Gems/Terrain/Code/Source/Components/TerrainSurfaceDataSystemComponent.cpp