This includes a few different optimizations found while trying to make the bulk query APIs faster:
* Switches mutexes over to shared_lock to optimize for the multi-reader-single-writer pattern
* Surface provider point creation now uses a pre-created set of masks to initialize with, and uses std::move() to move the created point into the output list instead of copying it.
* Splits CombineSortAndFilterNeightboringPoints so that the FilterPoints() can occur separately and efficiently with erase/remove_if, and avoids making a copy of the output points.
* Optimized SurfaceDataShapeComponent::ModifySurfacePoints
* Fixed potential bug where the sort wasn't stable since it only compared the Z value, and could have produced unexpected results for differing points with the exact same Z value.
* Fixed up a couple small bugs and missing checks in the unit tests

Before:
```
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
SurfaceDataBenchmark/BM_GetSurfacePoints/1024                 1649 ms         1656 ms            1
SurfaceDataBenchmark/BM_GetSurfacePoints/2048                 6453 ms         6453 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/1024       2215 ms         2219 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/2048       8692 ms         8688 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/1024         1962 ms         1969 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/2048         7966 ms         7969 ms            1
```
After:
```
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
SurfaceDataBenchmark/BM_GetSurfacePoints/1024                 1335 ms         1328 ms            1
SurfaceDataBenchmark/BM_GetSurfacePoints/2048                 5193 ms         5188 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/1024       2047 ms         2047 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/2048       7989 ms         7984 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/1024         1745 ms         1750 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/2048         7097 ms         7094 ms            1
```